pyproject.toml
[tool.poetry]
name = "bob"
description = "Sends voide messages"
version = "1.0.0"
authors = ["OpenAlcoholics <team@openalcoholics.group>"]
license = "MIT"

[tool.poetry.dependencies]
aiofiles = "^22.1.0"
certifi = "2022.12.7"
click = "^8.1.3"
Deprecated = "^1.2.13"
injector = "^0.20.1"
more-itertools = "^9.0.0"
pendulum = "^2.0.0"
python = "~3.11"
python-dotenv = "^0.21.0"
python-telegram-bot-raw = "20.0b0"
requests = "2.28.1"
sentry-sdk = "1.12.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "0.991"
pytest = "^7.1.2"
types-aiofiles = "^22.1.0"
types-Deprecated = "^1.2.9"
types-requests = "2.28.11.5"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true
